config/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "one-ai-config"
version = "0.1.0"
description = "Configuration schema, validator, and code generator for OpenNebula AI Assistant"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/one_ai_config"]

[tool.hatch.build.targets.wheel.force-include]
"src/one_ai_config/codegen/templates" = "one_ai_config/codegen/templates"

[tool.ruff]
line-length = 100
target-version = "py310"
